Address comments

+++ cruise-control-core/src/main/java/com/linkedin/cruisecontrol/monitor/sampling/aggregator/MetricValues.java
@@ -155,14 +155,14 @@ public class MetricValues {
   /**
    * @return the average value of all the values in this MetricValues.
    */
-  public double avg() {
+  public float avg() {
     return (float) (_sumForAvg / _values.length);
   }
 
   /**
    * @return the max value of all the values in this MetricValues.
    */
-  public double max() {
+  public float max() {
     if (_max >= 0) {
       return _max;
     } else {
